fix: Fix startup warnings #2 and #3 (menupack and startup script) (#389)
* fix: Fix startup warnings #2 and #3

Bug #2 (menupack lnumcursor corruption):
- Fixed v7 format path in meloadmenurecord_internal() to read into
  tysavedmenuinfo_v7 (1056 bytes) and convert, instead of incorrectly
  reading into tysavedmenuinfo which caused field misalignment
- Added proper byte order conversion using disk_to_host_* helpers
- Expanded tysavedmenuinfo.lnumcursor from short (16-bit) to int64_t
  for consistency with v7 disk format - avoids future confusion
- Removed loword() truncation in mesavemenurecord packing path
- Reduced waste[] padding from 42 to 36 bytes to account for expansion

Bug #3 (startup script warning):
- Removed misleading warning log that fired when bserror was populated
- The error buffer captures caught try/catch errors, not actual failures
- Script still completes successfully even when bserror has content
- Warning was cosmetic noise that confused users


* refactor: Expand line number types from long to int64_t

Standardize all in-memory cursor and line number types to int64_t for
consistency with v7 disk format. This is a mechanical change that:

- Updates struct fields: lnumbarcursor, lnumcursor, vertcurrent
- Updates function signatures in headers and implementations
- Updates local variables in outline/display/scroll code

Files changed: 22 (headers + implementations + test stubs)

Disk formats remain unchanged - pack/unpack code handles conversion
between 64-bit memory format and disk formats (which use split-word
pairs for backward compatibility).


* chore: Add _Static_assert for tysavedmenuinfo struct size

Adds compile-time verification that the in-memory menu struct remains
exactly 116 bytes. This prevents accidental size changes that would
break compatibility with tyOLD42savedmenuinfo and disk format layouts.

Addresses PR reviewer concern about struct packing verification.


* docs: Add comments and defensive checks addressing PR review feedback

- Add overflow validation for flags and menuactivelayer (fail on corrupt data)
- Add _Static_assert for tysavedmenuinfo_v7 size (1056 bytes)
- Use PRIu32/PRIx64/PRId64 format macros for portable logging
- Clarify comment about GUI-only fields (applies to both headless and desktop)


---------

+++ Common/source/oplineheight.c
@@ -81,21 +81,21 @@ hdlheadrecord opgetlastvisiblenode (void) {
 	} /*opgetlastvisiblenode*/
 
 
-long opgetcurrentscreenlines (boolean flscrollwise) {
-	
+int64_t opgetcurrentscreenlines (boolean flscrollwise) {
+
 	/*
 	return the number of lines currently showing in the window.
-	
+
 	6.0a12 dmb: added flscrollwise parameter. If true, return the number
-	of scroll lines on the screen; otherwise return the number of headlines, 
+	of scroll lines on the screen; otherwise return the number of headlines,
 	the original meaning if this function
 	*/
-	
+
 	hdloutlinerecord ho = op_get_outlinedata();
 	hdlheadrecord nomad = (**ho).hline1, nextnomad;
 	Rect r = (**ho).outlinerect;
-	long vertpixels = r.bottom - r.top;
-	short ctpixels = 0, ctlines = 0;
+	int64_t vertpixels = r.bottom - r.top;
+	int64_t ctpixels = 0, ctlines = 0;
 	
 	ctpixels += opgetline1top ();
 	
@@ -138,21 +138,21 @@ long opgetcurrentscreenlines (boolean flscrollwise) {
 	} /*opgetcurrentscreenlines*/
 	
 	
-long opsumprevlineheights (long lnum, short *heightthisline) {
-	
+int64_t opsumprevlineheights (int64_t lnum, short *heightthisline) {
+
 	/*
 	return the sum of the lineheights of all lines above this one.
-	
+
 	lnum is 0-based.
-	
+
 	6.0b2 dmb: work with negative lnums, with the same semantics (return
 	a negative number)
 	*/
-	
+
 	hdloutlinerecord ho = op_get_outlinedata();
 	hdlheadrecord nomad = (**ho).hline1;
-	long sum = 0;
-	long i;
+	int64_t sum = 0;
+	int64_t i;
 	
 	sum += opgetline1top ();
 	
@@ -183,10 +183,10 @@ long opsumprevlineheights (long lnum, short *heightthisline) {
 	} /*opsumprevlineheights*/
 	
 	
-long opsumalllineheights (void) {
-	
+int64_t opsumalllineheights (void) {
+
 	hdlheadrecord nomad = (**op_get_outlinedata()).hsummit, nextnomad;
-	long sum = 0;
+	int64_t sum = 0;
 	
 	while (true) {
 		
@@ -202,26 +202,26 @@ long opsumalllineheights (void) {
 	} /*opsumalllineheights*/
 	
 	
-long opgetlinestoscrollupforvisi (hdlheadrecord hnode) {
-	
+int64_t opgetlinestoscrollupforvisi (hdlheadrecord hnode) {
+
 	/*
 	return the number of lines you have to scroll to make the node
 	visible. assume the node lies off the bottom of the window.
-	
+
 	first we figure out how many pixels off we are (ctneeded).
-	
+
 	then we loop up from the first line until we equal or exceed
 	that number.
 	*/
-	
+
 	hdloutlinerecord ho = op_get_outlinedata();
 	hdlheadrecord nomad = (**ho).hline1;
 	Rect r = (**ho).outlinerect;
-	short vertpixels = r.bottom - r.top;
-	long ctpixels = 0;
-	long ctneeded = 0;
-	long ctlines;
-	long ctscroll;
+	int64_t vertpixels = r.bottom - r.top;
+	int64_t ctpixels = 0;
+	int64_t ctneeded = 0;
+	int64_t ctlines;
+	int64_t ctscroll;
 	short lh;	
 	
 	ctpixels = opgetline1top (); //measure  from top of line1
@@ -275,24 +275,24 @@ long opgetlinestoscrollupforvisi (hdlheadrecord hnode) {
 	} /*opgetlinestoscrollupforvisi*/
 
 
-long opgetlinestoscrolldownforvisi (hdlheadrecord hnode) {
-	
+int64_t opgetlinestoscrolldownforvisi (hdlheadrecord hnode) {
+
 	/*
 	return the number of lines you have to scroll to make the node
 	visible. assume the node lies above the top of the window.
-	
+
 	we count the number of scroll lines above hline1, up to and including
 	hnode
-	
+
 	6.0b2 dmb: account for text selection in headlines taller than the screen
 	*/
-	
+
 	hdloutlinerecord ho = op_get_outlinedata();
 	hdlheadrecord nomad = (**ho).hline1;
 	hdlheadrecord hsummit = (**ho).hsummit;
-	long defaultlineheight = (**ho).defaultlineheight;
-	long ctscroll = 0;
-	long ctpixels = 0;
+	int64_t defaultlineheight = (**ho).defaultlineheight;
+	int64_t ctscroll = 0;
+	int64_t ctpixels = 0;
 	Point ptsel;
 	Rect r;
